Risikomanagement und Erstmeldung - Updates

+++ backend/src/main/java/org/thomcgn/backend/falloeffnungen/risk/dto/RiskSnapshotResponse.java
@@ -8,8 +8,8 @@ public record RiskSnapshotResponse(
         Long falloeffnungId,
         Long configId,
         String configVersion,
-        double rawScore,
-        double protectiveReduction,
+        BigDecimal rawScore,
+        BigDecimal protectiveReduction,
         BigDecimal finalScore,
         String trafficLight,       // GRUEN|GELB|ROT
         String rationaleJson,      // JSON-String (Frontend kann darstellen)
